return validation of deleting session

+++ app/Http/Controllers/TrainingSessionController.php
@@ -57,11 +57,9 @@ class TrainingSessionController extends Controller
      */
     public function store(StoreTrainingSessionRequest $request)
     {
-        // $validated = $request->validated();
         $trainSessionWithCoach=array_slice(request()->all(),count(request()->all())-1);
         $trainingSession=array_slice(request()->all(),1,count(request()->all())-2);
         $trainingSession=TrainingSession::create($trainingSession);
-        
         $trainSessionWithCoach['training_session_id']=$trainingSession->id;
         DB::table('coach_training_session')->insert($trainSessionWithCoach);
         return view('menu.training_sessions.index');
@@ -128,17 +126,18 @@ class TrainingSessionController extends Controller
      */
     public function destroy($id)
     {
-        // $trainingSession = TrainingSession::find($id);
-        // if (
-        //     $trainingSession->strats_at < now() &&
-        //     $trainingSession->finishes_at > now() &&
-        //     $trainingSession->gym_members->count() > 0
-        // ) {
-        //     return 'Hahaha';
-        // } else {
+        $trainingSession = TrainingSession::find($id);
+        if (
+            $trainingSession->strats_at < now() &&
+            $trainingSession->finishes_at > now() &&
+            $trainingSession->gym_members->count() > 0
+        ) 
+        {
+            return response()->json(['fail' => 'Can\'t delete this session']);
+        } else {
             TrainingSession::find($id)->delete();
             return response()->json(['success' => 'This session has been deleted successfully']);
-        // }
+        }
     }
     public function getCoaches($gymId)
     {
